requirements.txt
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
unstructured
pinecone-client
huggingface_hub
sentence_transformers
llama-cpp-python
-e .